Adicionando classe DTO para indentificar o aluno
