--- MyApp/src/main/java/br/com/myApp/MyApp/model/enumerations/PeriodoEnum.java
@@ -1,7 +1,0 @@
-package br.com.myApp.MyApp.model.enumerations;
-
-public enum PeriodoEnum {
-
-    MANHA, TARDE, NOITE, INTEGRAL, MEIO_PERIODO_I, MEIO_PERIODO_II
-
-}
