--- MyApp/src/main/java/br/com/myApp/MyApp/model/enumerations/TurmaEnum.java
@@ -1,7 +1,0 @@
-package br.com.myApp.MyApp.model.enumerations;
-
-public enum TurmaEnum {
-
-        TURMA_A, TURMA_B, TURMA_C, TURMA_D, TURMA_E
-
-}
